% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_multipl_df.R
\name{load_multipl_df}
\alias{load_multipl_df}
\title{Load Multiple data set in R.}
\usage{
load_multipl_df(data_folder, file_format = ".rds", ...)
}
\arguments{
\item{data_folder}{a folder where Data is stored.}

\item{file_format}{file format ie (.csv, .dta, .rds, .xlsx)}

\item{...}{any other function}
}
\description{
The function help to load multiple data set from a specified folder within the working directory and save the results as a list.
Specifying the folder location works with here package. which help to obtain current working directory
}
