% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_Rmd_to_Rscript.R
\name{mult_Rmd_to_Rscript}
\alias{mult_Rmd_to_Rscript}
\title{Converts one or more multiple Rmarkdown into R scripts.}
\usage{
mult_Rmd_to_Rscript(Rmd_input_folder, Rscript_output_folder, ...)
}
\arguments{
\item{Rmd_input_folder}{a folder where Rmarkdown scripts are stored.}

\item{Rscript_output_folder}{a folder where Produced R-scripts should be stored.}

\item{...}{any other function}
}
\description{
The function help to convert multiple Rmarkdown into R scripts and save the results of the scripts in the folder given.
Specifying the folder location works with here package. which help to obtain current working directory
}
