% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_join.R
\name{multiple_join}
\alias{multiple_join}
\title{Join multiple data sets.}
\usage{
multiple_join(df_list, type_join = "left_join", ...)
}
\arguments{
\item{df_list}{a list of all data sets.}

\item{type_join}{type of join i.e ("left_join","right_join","full_join" & "anti_join").}

\item{...}{any other function}
}
\description{
The function help to Join multiple data set stored in a list. It perform four types of dplyr joins left_join, right_join,
full_join and anti_join. By default, the join type is left_join.
}
