% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Date.R
\name{parse_Date}
\alias{parse_Date}
\title{Converts POSIXct format to date}
\usage{
parse_Date(x, format = "mdy")
}
\arguments{
\item{x}{variable}

\item{format}{- date format i.e.only supports (mdy,ymd,dmy), default is mdy}
}
\description{
A function helps to convert Posixct date to normal dates. It also support string date.
}
