% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_rscripts.R
\name{run_multiple_rscripts}
\alias{run_multiple_rscripts}
\title{Opens and Runs one/ multiple External R-Scripts on the current script.}
\usage{
run_multiple_rscripts(rscript_folder, echo = F, ...)
}
\arguments{
\item{rscript_folder}{a folder where R-scripts are stored.}

\item{echo}{echo the external scripts when executing.}

\item{...}{any other function}
}
\description{
The function help to execute/run one or multiple R-scripts from a specified folder in the current working directory.
}
\details{
By default echo is False(F), but can be set to true(F) to see each line of code in the external script execute.
}
