% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified Systematic sampling-Remmap.R
\name{strat_syst_sampling}
\alias{strat_syst_sampling}
\title{Stratified Systematic Sample for sampling AIM 3 subjects from women who has gone through the screening process described in Aim 1  in REMAPP}
\usage{
strat_syst_sampling(
  data,
  k = 2,
  enrollment_month = "April",
  enrollment_year = 2023
)
}
\arguments{
\item{data}{- The REMAPP data set to be operated on}

\item{k}{- sample interval}

\item{enrollment_month}{- Enrollment month(s) e.g c("April","May)}

\item{enrollment_year}{- Enrollment Year e.g 2023}
}
\description{
Function allows you to sample AIM3 cohort for each trimester cross-sectional
}
