% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Systematic sampling.R
\name{syst_sampling}
\alias{syst_sampling}
\title{Systematic Sample function}
\usage{
syst_sampling(N, n)
}
\arguments{
\item{N}{- The data set to be operated on}

\item{n}{- sample size}
}
\description{
Function allows you to sample data based on Systematic sample.
}
